refactor: AddCard 컴포넌트 추가로인한 컴포넌트 구조 리팩토링
