+++ src/components/AddCard/NewCardForm/CardCVCInput/index.styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { FONT_SIZE } from '../../../constants/constant';
+import { FONT_SIZE } from '../../../../constants/constant';
 
 export const CardCVCInputWrapper = styled.div`
   margin-bottom: 10px;
