+++ src/components/AddCard/NewCardForm/CardPasswordInput/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { CardPasswordInputWrapper } from './index.styles';
-import Input from '../../../common/Input';
-import { INPUT } from '../../../constants/constant';
+import Input from '../../../../common/Input';
+import { INPUT } from '../../../../constants/constant';
 
 const CardPasswordInput = ({ password, errorMessage, onChangeCardInput }) => {
   return (
